Реализует получение данных с сервера
