+++ package.json
@@ -14,6 +14,7 @@
     "@babel/core": "7.10.2",
     "@babel/preset-env": "7.10.2",
     "@babel/preset-react": "7.10.1",
+    "axios-mock-adapter": "1.18.1",
     "babel-jest": "26.0.1",
     "babel-loader": "8.1.0",
     "enzyme": "3.11.0",
@@ -40,6 +41,7 @@
     "react-redux": "7.2.0",
     "react-router-dom": "5.2.0",
     "redux": "4.0.5",
+    "redux-devtools-extension": "2.13.8",
     "redux-thunk": "2.3.0",
     "reselect": "4.0.0"
   }
