+++ src/components/cities-list/cities-list.test.js
@@ -14,19 +14,21 @@ const currentCity = {
   cityCoords: [48.85341, 2.3488],
 };
 
+describe(`CitiesList`, () => {
+  it(`Should CitiesList render correctly`, () => {
+    const tree = renderer
+      .create(
+          <CitiesList
+            cities={cities}
+            currentCity={currentCity}
+            onCityClick={() => {}}
+          />,
+          {createNodeMock: () => {
+            return document.createElement(`div`);
+          }})
+      .toJSON();
 
-it(`Should MainScreen render correctly`, () => {
-  const tree = renderer
-    .create(
-        <CitiesList
-          cities={cities}
-          currentCity={currentCity}
-          onCityClick={() => {}}
-        />,
-        {createNodeMock: () => {
-          return document.createElement(`div`);
-        }})
-    .toJSON();
-
-  expect(tree).toMatchSnapshot();
+    expect(tree).toMatchSnapshot();
+  });
 });
+
