+++ src/components/offer-card/offer-card.jsx
@@ -4,23 +4,26 @@ import {OfferType} from "../../const.js";
 
 const OfferCard = (props) => {
   const {offer, onCardMouseEnter, onCardMouseLeave, onHeaderClick} = props;
+  const {isItPremium, img, price, rating, title, type, isItFavorite} = offer;
 
   return <article className="cities__place-card place-card" onMouseEnter={() => onCardMouseEnter(offer)} onMouseLeave={() => onCardMouseLeave()}>
-    {offer.isItPremium ? <div className="place-card__mark">
+    {isItPremium ? <div className="place-card__mark">
       <span>Premium</span>
     </div> : ``}
     <div className="cities__image-wrapper place-card__image-wrapper">
       <a href="#">
-        <img className="place-card__image" src={offer.img} width="260" height="200" alt="Place image" />
+        <img className="place-card__image" src={img} width="260" height="200" alt="Place image" />
       </a>
     </div>
     <div className="place-card__info">
       <div className="place-card__price-wrapper">
         <div className="place-card__price">
-          <b className="place-card__price-value">&euro;{offer.price}</b>
+          <b className="place-card__price-value">&euro;{price}</b>
           <span className="place-card__price-text">&#47;&nbsp;night</span>
         </div>
-        <button className="place-card__bookmark-button place-card__bookmark-button--active button" type="button">
+        <button className={isItFavorite ?
+          `place-card__bookmark-button place-card__bookmark-button--active button`
+          : `place-card__bookmark-button button`} type="button">
           <svg className="place-card__bookmark-icon" width="18" height="19">
             <use xlinkHref="#icon-bookmark"></use>
           </svg>
@@ -29,7 +32,7 @@ const OfferCard = (props) => {
       </div>
       <div className="place-card__rating rating">
         <div className="place-card__stars rating__stars">
-          <span style={{width: `${offer.rating * 20}%`}}></span>
+          <span style={{width: `${rating * 20}%`}}></span>
           <span className="visually-hidden">Rating</span>
         </div>
       </div>
@@ -37,9 +40,9 @@ const OfferCard = (props) => {
         <a href="#" onClick={(evt) => {
           evt.preventDefault();
           onHeaderClick(offer);
-        }}>{offer.title}</a>
+        }}>{title}</a>
       </h2>
-      <p className="place-card__type">{offer.type}</p>
+      <p className="place-card__type">{type}</p>
     </div>
   </article>;
 };
@@ -48,7 +51,7 @@ OfferCard.propTypes = {
   onCardMouseEnter: PropTypes.func.isRequired,
   onCardMouseLeave: PropTypes.func.isRequired,
   onHeaderClick: PropTypes.func.isRequired,
-  offer: OfferType.isRequired,
+  offer: OfferType,
 };
 
 export default OfferCard;
