+++ src/components/offers-list/offers-list.jsx
@@ -23,7 +23,7 @@ const OffersList = (props) => {
 
 OffersList.propTypes = {
   offers: PropTypes.arrayOf(
-      OfferType.isRequired
+      OfferType
   ).isRequired,
   onHeaderClick: PropTypes.func.isRequired,
   onCardMouseEnter: PropTypes.func.isRequired,
