+++ src/components/sort/sort.jsx
@@ -1,10 +1,11 @@
 import React, {PureComponent} from 'react';
 import {connect} from "react-redux";
 import PropTypes from "prop-types";
-
-import {ActionCreator} from "../../reducer/ui/ui.js";
+import {ActionCreator as ActionCreatorData} from "../../reducer/data/data.js";
+import {ActionCreator as ActionCreatorUi} from "../../reducer/ui/ui.js";
 import {SortType} from '../../const.js';
-import NameSpace from '../../reducer/name-space';
+import {getCurrentSortType} from '../../reducer/data/selectors.js';
+import {getShowSortMenu} from '../../reducer/ui/selectors.js';
 
 const menuItems = [
   {
@@ -77,16 +78,16 @@ class Sort extends PureComponent {
 
 const mapDispatchToProps = (dispatch) => ({
   onSort(sortType) {
-    dispatch(ActionCreator.sortOffers(sortType));
+    dispatch(ActionCreatorData.sortOffers(sortType));
   },
   onShowSortMenu(isOpened) {
-    dispatch(ActionCreator.showSortMenu(isOpened));
+    dispatch(ActionCreatorUi.showSortMenu(isOpened));
   }
 });
 
 const mapStateToProps = (state) => ({
-  currentSortType: state[NameSpace.UI].currentSortType,
-  isOpened: state[NameSpace.UI].showSortMenu,
+  currentSortType: getCurrentSortType(state),
+  isOpened: getShowSortMenu(state),
 });
 
 Sort.propTypes = {
