--- src/mock/cities.js
@@ -1,28 +1,0 @@
-const cities = [
-  {
-    cityName: `Paris`,
-    cityCoords: [48.85341, 2.3488],
-  },
-  {
-    cityName: `Cologne`,
-    cityCoords: [50.9381, 6.95],
-  },
-  {
-    cityName: `Brussels`,
-    cityCoords: [50, 8504, 4, 34878],
-  },
-  {
-    cityName: `Amsterdam`,
-    cityCoords: [52.38333, 4.9],
-  },
-  {
-    cityName: `Hamburg`,
-    cityCoords: [53.5753, 10.0153],
-  },
-  {
-    cityName: `Dusseldorf`,
-    cityCoords: [51.2217, 6.77616],
-  },
-];
-
-export default cities;
