+++ src/reducer/data/data.js
@@ -1,12 +1,13 @@
-import {getOffersInCity, extend} from '../../utils';
-import {adaptOffersAll} from '../../mock/offers';
-// import offers from '../../mock/offers';
+import {getOffersInCity, extend, getSortedOffers} from '../../utils';
+import {adaptOffersAll} from '../../adapter/offers';
+import {SortType} from '../../const.js';
 
 const initialState = {
   city: {
     cityName: `Amsterdam`,
     cityCoords: [52.38333, 4.9],
   },
+  currentSortType: `Popular`,
   allOffers: [],
   currentOffers: [],
 };
@@ -15,6 +16,7 @@ const ActionType = {
   LOAD_OFFERS: `LOAD_OFFERS`,
   CHANGE_LOCATION: `CHANGE_LOCATION`,
   CHANGE_CURRENT_OFFERS: `CHANGE_CURRENT_OFFERS`,
+  SORT_OFFERS: `SORT_OFFERS`,
 };
 
 const ActionCreator = {
@@ -32,6 +34,10 @@ const ActionCreator = {
     type: ActionType.CHANGE_CURRENT_OFFERS,
     payload: cityName,
   }),
+  sortOffers: (sortType) => ({
+    type: ActionType.SORT_OFFERS,
+    payload: sortType,
+  }),
 };
 
 const Operation = {
@@ -46,12 +52,21 @@ const Operation = {
 const reducer = (state = initialState, action) => {
   switch (action.type) {
     case ActionType.LOAD_OFFERS:
-      return extend(state, {offersAll: action.payload});
+      return extend(state, {allOffers: action.payload,
+        currentOffers: getOffersInCity(state.city.cityName, action.payload)}
+      );
     case ActionType.CHANGE_LOCATION:
       return extend(state, {city: action.payload});
     case ActionType.CHANGE_CURRENT_OFFERS:
       return extend(state, {
         currentOffers: getOffersInCity(action.payload, state.allOffers),
+      });
+    case ActionType.SORT_OFFERS:
+      return extend(state, {
+        currentSortType: action.payload,
+        allOffers: action.payload === SortType.POPULAR
+          ? state.allOffers
+          : getSortedOffers(state.allOffers, action.payload),
       });
     default:
       return state;
