+++ src/reducer/data/selectors.js
@@ -1,17 +1,22 @@
 import {createSelector} from "reselect";
 import NameSpace from "../name-space.js";
 
+const NAME_SPACE = NameSpace.DATA;
 
 export const getOffers = (state) => {
-  return state[NameSpace.DATA].offers;
+  return state[NAME_SPACE].allOffers;
+};
+
+export const getCurrentSortType = (state) => {
+  return state[NAME_SPACE].currentSortType;
 };
 
 export const getCurrentCity = (state) => {
-  return state[NameSpace.DATA].currentCity;
+  return state[NAME_SPACE].city;
 };
 
 export const getCurrentCityName = (state) => {
-  return state[NameSpace.DATA].currentCity.cityName;
+  return state[NAME_SPACE].city.cityName;
 };
 
 export const getCurrentOffers = createSelector(
