+++ src/reducer/reducer.js
@@ -1,8 +1,8 @@
 import {combineReducers} from "redux";
-import {reducer as data} from "./reducer/data/data.js";
-import {reducer as ui} from "./reducer/ui/ui.js";
+import {reducer as data} from "./data/data.js";
+import {reducer as ui} from "./ui/ui.js";
 // import {reducer as user} from "./reducer/user/user.js";
-import NameSpace from "./reducer/name-space.js";
+import NameSpace from "./name-space.js";
 
 export default combineReducers({
   [NameSpace.DATA]: data,
