+++ src/reducer/ui/selectors.js
@@ -3,10 +3,6 @@ import NameSpace from "../name-space.js";
 
 const NAME_SPACE = NameSpace.UI;
 
-export const getCurrentSortType = (state) => {
-  return state[NAME_SPACE].currentSortType;
-};
-
 export const getActiveOfferId = (state) => {
   return state[NAME_SPACE].activeOfferId;
 };
